Actually sort the records, try to preserve order
What this will actually do is sort by date and then by category. This is
because of how the records are read -- all of a category is read, and
then if the date is the same as other it will not reorder. This will
keep all of the tasks of some category completed on the same day
together.

I sort of would rather like to just see it as it happens, but there's
actually no way to define that since they are in separate folders,
unless I keep more fine-grained timestamps. I might consider doing this.