-t with -l only lists tasks that are due that day
This is to make it a bit easier to see what is coming down the pipeline.

Note that this doesn't actually consider "overdue" tasks or tasks that
have been delayed over multiple days with -n. I might change this at
some point, since it's not an accurate representation of what I need to
do, e.g. "todo -t today -l" is not the same as "todo -l".